改进

+++ 至美微博/至美微博/Classes/Others/Category/NSString+Size.h
@@ -14,6 +14,6 @@
 +(CGSize)sizeWithText:(NSString*)text maxSize:(CGSize)maxSize font:(UIFont*)font;
 
 //计算文件或目录大小
--(long long)fileSizeWithFilePath;
+-(CGFloat)fileSizeWithFilePath;
 
 @end
